Problem Set 2
Hangman: a program for a hangman game for 1 player.  Gives the user 10
guesses, show all the letters left for the player to guess, and shows
the displayed word with the correct guess.
Newton: Program has a function that finds the derivative of a polynomial
(in the form of a list), and the solution to a polynomial.  The final
part uses the above functions to find a root of a polynomial with
newton's method.